Card index working

+++ db/migrate/20150515180009_add_color_costs_to_cards.rb
@@ -1,11 +1,11 @@
 class AddColorCostsToCards < ActiveRecord::Migration
   def change
     change_table :cards do |t|
-      t.decimal :cost_white
-      t.decimal :cost_blue
-      t.decimal :cost_green
-      t.decimal :cost_red
-      t.decimal :cost_brown
+      t.integer :cost_white
+      t.integer :cost_blue
+      t.integer :cost_green
+      t.integer :cost_red
+      t.integer :cost_brown
     end
   end
 end
